Allow autocompletion of executables in ./bin
Without an explicit, full path to the current ./bin, zsh's
autocompletion doesn't know what to do with `bin` in $PATH. This adds a
chpwd hook to update $PATH to include the new directory's `bin`
directory. This is a bit brute-force-y, and depends on us _always_
having an entry for the current directory (meaning, if we remove
duplicates, this breaks in $HOME).