Set up Homebrew shellenv portably
If Homebrew is installed, we want to set things like PATH to include
Homebrew. On Intel, this mostly happens by magic (because macOS includes
/usr/local/bin, where Homebrew installs things), but on Apple Silicon
it doesn't work by default (because Homebrew on Apple Silicon installs
to /opt/homebrew).

This checks if `brew` is in either of the expected locations, then uses
it to set the shell variables appropriately.